Removed GameState.NBTL and GameState.DTTL properties; these are now dynamically calculated as required.  The Engine.CheckEnemies method has been removed as well, since it is no longer needed.  Along with the deprecated GameState.Wt, this removes a large source of potentially tricky bugs when building new games.
The only caveat remaining:  for any Monster record with a Friendliness percent (100-200), the ResolveFriendlinessPct method must be called to resolve it to a member of Enum.Friendliness before it is used during game play.  This is always done at game startup by InitMonsters, but if you need to reset the Friendliness percent to recalculate the Monster's Friendliness, you must call the method explicitly.  See the death dog's GiveCommand logic in Wrenhold's Secret Vigil for an example of this.