Bug fix:  LightCommand now omits display of unseen artifact Descs when in an unlit room.
*** NOTE TO ALL REPOSITORY USERS ***

The Eamon-CS and Eamon-CS-Misc GitHub repositories have become inflated over the last few years with numerous binary blobs checked in (.dll, .pdb, .pdf, .apk, .zip files, etc).  After some experimenting, I successfully created a private Eamon-CS repo with the excess cruft purged and all the valuable history intact.  This repo is on the order of 80% smaller.

The bad news is history is being rewritten which means, your old repository will be incompatible with the new one, and attempts to Fetch will fail; you will have to re-clone the repository to get back in sync with the one on GitHub.

Both repositories will be rebuilt sometime in late June or early July, after Eamon CS Mobile 1.6 is released.  The README.md will be updated as a reminder prior to this occurring.  Also, you can expect repo rebuilds to happen once in a while on a go forward basis as needed (perhaps once a year).

Apologies for any inconveniences this may cause.